fix:샘플데이터data 폴더안으로 코드 변경
