+++ src/components/community/PostDetail.tsx
@@ -39,35 +39,7 @@ const PostDetail = () => {
     // 실제 삭제 API 연결 예정
     console.log('게시글 삭제:', post.id);
   };
-
-  return (
-    <div className="w-full max-w-[800px] mx-auto px-4 sm:px-6">
-      <CommunityTitle title="감정 소비 이야기" />
-
-      <div className="border bg-white rounded-lg p-[30px] shadow-sm">
-        {/* 작성자/작성일/더보기 */}
-        <div className="flex items-center justify-between mb-5">
-          <div className="flex items-center gap-2">
-            <AuthorInfo author={author} />
-            <span className="text-xs text-primary-500 ml-2">{formattedDate}</span>
-          </div>
-          {isMine && (
-            <div className="relative">
-              <PostMoreButton postData={post} onEdit={handleEdit} onDelete={handleDelete} />
-            </div>
-          )}
-        </div>
-
-        {/* 제목 */}
-        <h2 className="text-xl font-semibold text-gray-800 mb-3">{title}</h2>
-
-    console.log('게시글 수정');
-  };
-
-  const handleDelete = () => {
-    console.log('게시글 삭제');
-  };
-
+  
   return (
     <div className="w-full max-w-[800px] mx-auto px-4 sm:px-6">
       <CommunityTitle title="감정 소비 이야기" />
