+++ src/types/wallet.ts
@@ -11,7 +11,7 @@ export type Category = ExpenseCategory | IncomeCategory;
 
 // Wallet 타입 정의
 export type Wallet = {
-  id: number;
+  id: string;
   type: 'expense' | 'income';
   amount: number; // 금액
   title: string; // 제목
@@ -19,6 +19,30 @@ export type Wallet = {
   category: ExpenseCategory | IncomeCategory;
   emotion: Emotion; // 고정된 8개의 이모션 중 하나
   date: Date;
+};
+
+export type SWallet = {
+  walletUuid: string
+  title: string;
+  content?: string;
+  amount: number;
+  type: "INCOME" | "EXPENSE";
+  category: ExpenseCategory | IncomeCategory;
+  emotion: Emotion;
+  date: string;
+}
+
+export const transformSWalletToWallet = (s: SWallet): Wallet => {
+  return {
+    id: s.walletUuid,
+    title: s.title,
+    content: s.content,
+    amount: s.amount,
+    category: s.category,
+    emotion: s.emotion,
+    date: new Date(s.date),
+    type: s.type === "INCOME" ? "income" : "expense",
+  };
 };
 
 // 날짜별 WalletList 타입 정의
@@ -35,34 +59,3 @@ export type MonthlyWalletList = {
   totalExpense : number;
   entries: DailyWalletList[];
 }
-
-
-// 예시 데이터
-export const sampleData: MonthlyWalletList = {
-  year: 2025,
-  month: 5,
-  totalIncome: 4000000,
-  totalExpense: 404000,
-  entries: [
-    {
-      date: new Date("2025-05-01"),
-      totalAmount: 4000000,
-      entries: [{ id: 1, type: "income", amount: 4000000, category: "급여", title: "급여", emotion: "기대", date: new Date("2025-04-01") }],
-    },
-    {
-      date: new Date("2025-05-11"),
-      totalAmount: -300000,
-      entries: [{ id: 2, type: "expense", amount: -300000, category: "쇼핑", title: "쇼핑", emotion: "기대", date: new Date("2025-04-11") }],
-    },
-    {
-      date: new Date("2025-05-19"),
-      totalAmount: -12000,
-      entries: [{ id: 3, type: "expense", amount: -12000, category: "생활", title: "커피", emotion: "위로", date: new Date("2025-04-19") }],
-    },
-    {
-      date: new Date("2025-05-24"),
-      totalAmount: -250000,
-      entries: [{ id: 4, type: "expense", amount: -250000, category: "식비", title: "식비", emotion: "만족", date: new Date("2025-04-24") }],
-    },
-  ],
-};
